requirements.txt
ccxt
schedule
load_dotenv